pyproject.toml
[project]
name = "test-promocode"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django>=6.0.3",
    "django-injector>=0.3.1",
    "djangorestframework>=3.16.1",
    "drf-yasg>=1.21.15",
    "gunicorn>=25.1.0",
    "pytest>=9.0.2",
    "pytest-django>=4.12.0",
    "python-dotenv>=1.2.2",
]
